+++ aacgmv2/__init__.py
@@ -10,6 +10,9 @@ wrapper : Contains current python functions
 
 Parameters
 ---------------------------------------------------------------------------
+logger
+high_alt_coeff
+high_alt_trace
 AACGM_V2_DAT_PREFIX
 IGRF_12_COEFFS
 _aacgmv2.G2A
@@ -30,6 +33,8 @@ get_aacgm_coord_arr
 convert
 convert_mlt
 wrapper.set_coeff_path
+wrapper.test_height
+wrapper.test_time
 deprecated.subsol
 _aacgmv2.convert
 _aacgmv2.set_datetime
@@ -39,11 +44,31 @@ _aacgmv2.inv_mlt_convert
 _aacgmv2.inv_mlt_convert_yrsec
 ---------------------------------------------------------------------------
 """
+# Imports
+#---------------------------------------------------------------------
+
+import logging
 import os as _os
 from sys import stderr
-import logbook as logging
 
-__version__ = "2.5.1"
+from aacgmv2.wrapper import (convert_latlon, convert_mlt, get_aacgm_coord)
+from aacgmv2.wrapper import (convert_latlon_arr, get_aacgm_coord_arr)
+from aacgmv2.wrapper import (convert_bool_to_bit, convert_str_to_bit)
+from aacgmv2 import (deprecated)
+from aacgmv2.deprecated import (convert)
+from aacgmv2 import (_aacgmv2)
+
+# Define global variables
+#---------------------------------------------------------------------
+
+__version__ = "2.5.2"
+
+# Define a logger object to allow easier log handling
+logger = logging.getLogger('aacgmv2_logger')
+
+# Altitude constraints
+high_alt_coeff = 2000.0 # Tested and published in Shepherd (2014)
+high_alt_trace = 6378.0 # 1 RE, these are ionospheric coordinates
 
 # path and filename prefix for the IGRF coefficients
 AACGM_v2_DAT_PREFIX = _os.path.join(_os.path.realpath( \
@@ -73,23 +98,3 @@ _os.environ['AACGM_v2_DAT_PREFIX'] = AACGM_v2_DAT_PREFIX
 if __reset_warn__:
     stderr.write("non-default coefficient files may be specified by running " +
                  "aacgmv2.wrapper.set_coeff_path before any other functions\n")
-# Imports
-#---------------------------------------------------------------------
-
-try:
-    from aacgmv2.wrapper import (convert_latlon, convert_mlt, get_aacgm_coord)
-    from aacgmv2.wrapper import (convert_latlon_arr, get_aacgm_coord_arr)
-    from aacgmv2.wrapper import (convert_bool_to_bit, convert_str_to_bit)
-except Exception as err:
-    logging.exception(__file__ + ' -> aacgmv2: ' + str(err))
-
-try:
-    from aacgmv2 import (deprecated)
-    from aacgmv2.deprecated import (convert)
-except Exception as err:
-    logging.exception(__file__ + ' -> aacgmv2: ' + str(err))
-
-try:
-    from aacgmv2 import (_aacgmv2)
-except Exception as err:
-    logging.exception(__file__ + ' -> aacgmv2: ' + str(err))