+++ aacgmv2/tests/test_dep_aacgmv2.py
@@ -2,10 +2,13 @@
 from __future__ import division, absolute_import, unicode_literals
 
 import datetime as dt
+from io import StringIO
+import logging
 import numpy as np
 import pytest
-import aacgmv2
 import warnings
+
+import aacgmv2
 
 class TestFutureDepWarning:
     def setup(self):
@@ -74,7 +77,32 @@ class TestDepAACGMV2Warning(TestFutureDepWarning):
         self.test_routine = aacgmv2.deprecated.igrf_dipole_axis
         self.test_args = [self.dtime]
         self.test_future_dep_warning()
-        
+
+
+class TestDepLogging:
+    def setup(self):
+        """Runs before every method to create a clean testing setup"""
+
+        self.log_capture = StringIO()
+        aacgmv2.logger.addHandler(logging.StreamHandler(self.log_capture))
+
+        self.in_convert = ([60], [0], [-1], dt.datetime(2015, 1, 1, 0, 0, 0))
+        self.lwarn = u"conversion not intended for altitudes < 0 km"
+        self.lout = ''
+
+    def teardown(self):
+        """Runs after every method to clean up previous testing"""
+        self.log_capture.close()
+        del self.in_convert, self.lwarn, self.lout
+
+    def test_warning_below_ground_convert(self):
+        """ Test that a warning is issued if altitude is below zero"""
+
+        aacgmv2.convert(*self.in_convert)
+        self.lout = self.log_capture.getvalue()
+        assert self.lout.find(self.lwarn) >= 0
+
+
 class TestDepAACGMV2:
     def setup(self):
         """Runs before every method to create a clean testing setup"""
@@ -227,7 +255,7 @@ class TestDepAACGMV2:
 
     def test_convert_time_failure(self):
         """Test conversion with a bad time"""
-        with pytest.raises(AssertionError):
+        with pytest.raises(ValueError):
             with warnings.catch_warnings():
                 warnings.simplefilter("ignore")
                 self.lat, self.lon = aacgmv2.convert([60], [0], [300], None)
@@ -241,20 +269,6 @@ class TestDepAACGMV2:
         np.testing.assert_allclose(self.lat, [58.2258], rtol=1e-4)
         np.testing.assert_allclose(self.lon, [81.1685], rtol=1e-4)
 
-    def test_warning_below_ground_convert(self):
-        """ Test that a warning is issued if altitude is below zero"""
-        import logbook
-        lwarn = u"conversion not intended for altitudes < 0 km"
-
-        with logbook.TestHandler() as handler:
-            with warnings.catch_warnings():
-                warnings.simplefilter("ignore")
-                self.lat, self.lon = aacgmv2.convert([60], [0], [-1],
-                                                     self.dtime)
-                assert handler.has_warning(lwarn)
-
-        handler.close()
-
     def test_convert_maxalt_failure(self):
         """For an array, test failure for an altitude too high for
         coefficients"""
@@ -265,7 +279,7 @@ class TestDepAACGMV2:
 
     def test_convert_lat_failure(self):
         """Test error return for co-latitudes above 90 for an array"""
-        with pytest.raises(AssertionError):
+        with pytest.raises(ValueError):
             with warnings.catch_warnings():
                 warnings.simplefilter("ignore")
                 aacgmv2.convert([91, 60, -91], 0, 300, self.dtime)