+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 * Added FutureWarning to deprecated functions
 * Updated names in licenses
 * Moved module structure routine tests to their own class
+* Added high altitude limit to avoid while-loop hanging
+* Changed version support to 2.7, 3.6, and 3.7
+* Removed logbook dependency
 * Added logic to avoid reseting environment variables if not necessary